chore(suite): refactor banners folder
