+++ packages/suite/src/components/suite/banners/SuiteBanners/FirmwareHashMismatchBanner.tsx
@@ -1,8 +1,6 @@
-import { Banner } from './Banner';
+import { Banner } from '../Banner';
 import { Translation } from 'src/components/suite';
 
-const FirmwareHashMismatch = () => (
+export const FirmwareHashMismatch = () => (
     <Banner variant="critical" body={<Translation id="TR_FIRMWARE_HASH_MISMATCH" />} />
 );
-
-export default FirmwareHashMismatch;
