+++ packages/suite/src/components/suite/banners/SuiteBanners/OnlineStatusBanner.tsx
@@ -1,14 +1,12 @@
 import { Translation } from 'src/components/suite';
-import { Banner } from './Banner';
+import { Banner } from '../Banner';
 
 interface OnlineStatusProps {
     isOnline: boolean;
 }
 
-const OnlineStatus = ({ isOnline }: OnlineStatusProps) => {
+export const OnlineStatus = ({ isOnline }: OnlineStatusProps) => {
     if (isOnline) return null;
 
     return <Banner variant="critical" body={<Translation id="TR_YOU_WERE_DISCONNECTED_DOT" />} />;
 };
-
-export default OnlineStatus;
