+++ packages/suite/src/components/suite/banners/index.ts
@@ -1,0 +1,2 @@
+export { MessageSystemBanner } from './MessageSystemBanner';
+export { SuiteBanners } from './SuiteBanners/SuiteBanners';
